+++ app/src/main/res/layout/fragment_feed.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.dashboard.FeedFragment">
+        tools:context=".ui.feed.FeedFragment">
 
         <com.google.android.material.chip.ChipGroup
             android:id="@+id/cg_type"
